src/backend/requirements.txt
# Flask web framework (same version as Derik)
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0

# Database (same as Derik)
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# LangChain and AI (updated for with_structured_output support)
langchain==0.3.14
langchain-openai==0.2.14
langchain-community==0.3.14
openai>=1.6.1

# Data processing
pandas==2.1.4
openpyxl==3.1.2
numpy>=1.26.4

# Document processing (for AI role mapping feature)
PyPDF2==3.0.1
python-docx==1.2.0

# Vector database for RAG
faiss-cpu>=1.8.0

# Validation and utilities
pydantic>=2.7.4
python-dotenv==1.0.0
click==8.1.7

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.0.0

# Development tools
black==23.12.1
flake8==6.1.0
isort==5.13.2